Fix @channel / @channels ambiguity and issue #172
I removed the the @channel attribute because the behavior of the library
is confusing if someone specifies both the channel during initialization
and sets @channels later on.

This way we only use @channels which is an obvious array of @channels someone
wants to target and the channel given in the initialization ends up there.

I also fixed #172 by making it so that automatically the channels end up in
the where condition if @type is specified at the same time.